requirements-gpu.txt
# Requirements file for GPU usage 

# PyTorch and torchvision for GPU (CUDA 11.0)
# Install using: pip install torch==1.7.0+cu110 torchvision==0.8.1+cu110 -f https://download.pytorch.org/whl/torch_stable.html
# For CPU-only fallback, run requirementt-cpu.txt instead

torch==2.0.1+cu118
torchvision==0.15.2+cu118
tensorflow-gpu==2.8.0
tensorflow-addons==0.16.1
tensorflow-io==0.24.0

opencv-python==4.5.4.58
opencv-contrib-python==4.5.4.58

natsort==7.1.1

numpy==1.23.5
scipy==1.10.1

matplotlib==3.5.3
scikit-learn==1.2.2
scikit-image==0.19.2

wandb==0.12.11
tqdm==4.62.3
opencv-python==4.5.5.62

seaborn==0.11.2